static-site/builder/ui/requirements.txt
flask>=3.0.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
gevent>=23.9.0
Pillow>=10.0.0
